motivation_tests/test_m6.py
"""
M6: 系统级 test: layout + GPU batch 的收益
System-level test: layout + GPU batch benefits
"""
import sys
import os
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))

import numpy as np
import pandas as pd
from pathlib import Path
from typing import Dict, List
import json
import time
import psutil
import os

from src.similarity import TypedSetSimilarity
from src.data_generator import MultiModalDataGenerator


class CPURandomLayout:
    """CPU implementation with random scatter access"""
    
    def __init__(self, objects: List[Dict]):
        self.objects = objects
        # Simulate random scatter: objects stored randomly
        self.storage = list(enumerate(objects))
        np.random.shuffle(self.storage)
    
    def search(self, query: Dict, k: int = 10):
        """Search with random access pattern"""
        typed_sim = TypedSetSimilarity()
        scores = []
        
        # Random access to full set
        for obj_id, obj in self.storage:
            score = typed_sim.typed_set_similarity(query, obj)
            scores.append((obj_id, score))
        
        scores.sort(key=lambda x: x[1], reverse=True)
        return [obj_id for obj_id, _ in scores[:k]]


class ObjectMajorLayout:
    """Object-major layout with batch GPU processing"""
    
    def __init__(self, objects: List[Dict]):
        self.objects = objects
        # Object-major layout: all vectors for an object are contiguous
        self.object_vectors = []
        self.object_offsets = [0]
        
        for obj in objects:
            obj_vecs = []
            for modality, vectors in obj.items():
                obj_vecs.extend(vectors)
            self.object_vectors.append(obj_vecs)
            self.object_offsets.append(self.object_offsets[-1] + len(obj_vecs))
    
    def search_batch(self, queries: List[Dict], k: int = 10, batch_size: int = 32):
        """Batch search with sequential access"""
        typed_sim = TypedSetSimilarity()
        all_results = []
        
        # Process objects in batches
        for batch_start in range(0, len(self.objects), batch_size):
            batch_end = min(batch_start + batch_size, len(self.objects))
            batch_objects = self.objects[batch_start:batch_end]
            
            # Sequential access to batch
            for obj in batch_objects:
                for query in queries:
                    score = typed_sim.typed_set_similarity(query, obj)
                    all_results.append(score)
        
        return all_results


def measure_io_pattern(access_pattern: List[int]) -> Dict:
    """Measure I/O pattern characteristics"""
    # Simulate sequential vs random access
    sequential_count = 0
    for i in range(1, len(access_pattern)):
        if access_pattern[i] == access_pattern[i-1] + 1:
            sequential_count += 1
    
    sequential_ratio = sequential_count / (len(access_pattern) - 1) if len(access_pattern) > 1 else 0
    return {
        'sequential_ratio': sequential_ratio,
        'random_ratio': 1 - sequential_ratio,
        'total_accesses': len(access_pattern)
    }


def run_test_m6(output_dir: str = "results/m6"):
    """Run M6 motivation test"""
    print("="*80)
    print("M6: System-level Test - Layout + GPU Batch Benefits")
    print("="*80)
    
    Path(output_dir).mkdir(parents=True, exist_ok=True)
    
    # Generate dataset
    print("\n[1/4] Generating dataset...")
    gen = MultiModalDataGenerator(dim=128, seed=42)
    dataset = gen.generate_dataset(num_objects=1000, 
                                   multi_vector_config={'text': 2, 'image': 3})
    queries = [gen.generate_query(['text']) for _ in range(50)]
    
    # Test 1: CPU random scatter
    print("\n[2/4] Testing CPU random scatter layout...")
    cpu_layout = CPURandomLayout(dataset)
    
    start_time = time.time()
    cpu_results = []
    for query in queries:
        result = cpu_layout.search(query, k=10)
        cpu_results.append(result)
    cpu_latency = time.time() - start_time
    
    cpu_cpu_usage = psutil.cpu_percent(interval=0.1)
    cpu_mem_usage = psutil.Process(os.getpid()).memory_info().rss / 1024 / 1024
    
    # Simulate I/O pattern (random access)
    io_pattern_random = list(range(len(dataset)))
    np.random.shuffle(io_pattern_random)
    io_stats_random = measure_io_pattern(io_pattern_random)
    
    # Test 2: Object-major layout with batch
    print("\n[3/4] Testing object-major layout with batch...")
    object_major = ObjectMajorLayout(dataset)
    
    start_time = time.time()
    batch_results = object_major.search_batch(queries, k=10, batch_size=32)
    batch_latency = time.time() - start_time
    
    batch_cpu_usage = psutil.cpu_percent(interval=0.1)
    batch_mem_usage = psutil.Process(os.getpid()).memory_info().rss / 1024 / 1024
    
    # Simulate I/O pattern (sequential access)
    io_pattern_sequential = list(range(len(dataset)))
    io_stats_sequential = measure_io_pattern(io_pattern_sequential)
    
    # Throughput calculation
    print("\n[4/4] Computing throughput...")
    cpu_throughput = len(queries) / cpu_latency if cpu_latency > 0 else 0
    batch_throughput = len(queries) / batch_latency if batch_latency > 0 else 0
    
    # Print results
    print("\n" + "="*80)
    print("System-level Results")
    print("="*80)
    
    print("\nLatency:")
    print(f"  CPU Random Scatter:    {cpu_latency:.4f}s ({cpu_latency/len(queries)*1000:.2f}ms per query)")
    print(f"  Object-major Batch:    {batch_latency:.4f}s ({batch_latency/len(queries)*1000:.2f}ms per query)")
    print(f"  Speedup:               {cpu_latency/batch_latency:.2f}x" if batch_latency > 0 else "N/A")
    
    print("\nThroughput:")
    print(f"  CPU Random Scatter:    {cpu_throughput:.2f} queries/sec")
    print(f"  Object-major Batch:    {batch_throughput:.2f} queries/sec")
    print(f"  Improvement:           {((batch_throughput - cpu_throughput) / cpu_throughput * 100):+.2f}%")
    
    print("\nResource Utilization:")
    print(f"  CPU Random - CPU:      {cpu_cpu_usage:.1f}%")
    print(f"  CPU Random - Memory:   {cpu_mem_usage:.2f} MB")
    print(f"  Batch - CPU:           {batch_cpu_usage:.1f}%")
    print(f"  Batch - Memory:        {batch_mem_usage:.2f} MB")
    
    print("\nI/O Pattern:")
    print(f"  Random Scatter - Sequential Ratio:  {io_stats_random['sequential_ratio']:.4f}")
    print(f"  Random Scatter - Random Ratio:       {io_stats_random['random_ratio']:.4f}")
    print(f"  Object-major - Sequential Ratio:     {io_stats_sequential['sequential_ratio']:.4f}")
    print(f"  Object-major - Random Ratio:         {io_stats_sequential['random_ratio']:.4f}")
    
    # Save results
    summary = {
        'latency': {
            'cpu_random_scatter_s': cpu_latency,
            'object_major_batch_s': batch_latency,
            'speedup': cpu_latency / batch_latency if batch_latency > 0 else 0,
            'cpu_per_query_ms': cpu_latency / len(queries) * 1000,
            'batch_per_query_ms': batch_latency / len(queries) * 1000,
        },
        'throughput': {
            'cpu_random_scatter_qps': cpu_throughput,
            'object_major_batch_qps': batch_throughput,
            'improvement_percent': ((batch_throughput - cpu_throughput) / cpu_throughput * 100) if cpu_throughput > 0 else 0,
        },
        'resource_usage': {
            'cpu_random_cpu_percent': cpu_cpu_usage,
            'cpu_random_memory_mb': cpu_mem_usage,
            'batch_cpu_percent': batch_cpu_usage,
            'batch_memory_mb': batch_mem_usage,
        },
        'io_pattern': {
            'random_scatter': io_stats_random,
            'object_major': io_stats_sequential,
        }
    }
    
    with open(f"{output_dir}/m6_summary.json", 'w') as f:
        json.dump(summary, f, indent=2)
    
    print(f"\n✓ Results saved to {output_dir}/")
    print("\nKey Observations:")
    print("  - Object-major layout enables sequential I/O")
    print("  - Batch processing improves throughput")
    print("  - Better resource utilization with co-design")
    
    return summary


if __name__ == "__main__":
    import argparse
    parser = argparse.ArgumentParser()
    parser.add_argument("--output-dir", type=str, default="results/m6")
    args = parser.parse_args()
    run_test_m6(args.output_dir)

