requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
hnswlib>=0.7.0
torch>=2.0.0
Pillow>=10.0.0

